Finds the Smallest Value in an Array Using Multi-Threading
This program uses an insertion sort function that sorts only part of an
array. The program then creates
many threads that call this function to sort different sections of the
array. The result is that each partition is
sorted. Once the threads are done executing the main thread can then run
a function that uses these partitions as a list of potential minimum
values which it scans to find the smallest.  Windows32 and Posix
Versions of code done.